applied comments
