+++ cvat-ui/src/actions/settings-actions.ts
@@ -3,11 +3,17 @@
 //
 // SPDX-License-Identifier: MIT
 
+import _ from 'lodash';
 import { AnyAction } from 'redux';
+import { ThunkAction } from 'utils/redux';
 import {
     GridColor, ColorBy, SettingsState, ToolsBlockerState,
+    CombinedState,
 } from 'reducers';
 import { ImageFilter, ImageFilterAlias } from 'utils/image-processing';
+import { conflict, conflictDetector } from 'utils/conflict-detector';
+import GammaCorrection, { GammaFilterOptions } from 'utils/fabric-wrapper/gamma-correciton';
+import { shortcutsActions } from './shortcuts-actions';
 
 export enum SettingsActionTypes {
     SWITCH_ROTATE_ALL = 'SWITCH_ROTATE_ALL',
@@ -408,4 +414,111 @@ export function resetImageFilters(): AnyAction {
         type: SettingsActionTypes.RESET_IMAGE_FILTERS,
         payload: {},
     };
+}
+
+export function restoreSettingsAsync(): ThunkAction {
+    return async (dispatch, getState): Promise<void> => {
+        const state: CombinedState = getState();
+        const { settings, shortcuts } = state;
+
+        dispatch(shortcutsActions.setDefaultShortcuts(structuredClone(shortcuts.keyMap)));
+        const newSettings = { ..._.pick(settings, 'player', 'workspace'), imageFilters: [] as ImageFilter[] };
+        const settingsString = localStorage.getItem('clientSettings') as string;
+        if (!settingsString) return;
+        const loadedSettings = JSON.parse(settingsString);
+        for (const [sectionKey, section] of Object.entries(newSettings)) {
+            for (const [key, value] of Object.entries(section)) {
+                let settedValue = value;
+                if (sectionKey in loadedSettings && key in loadedSettings[sectionKey]) {
+                    settedValue = loadedSettings[sectionKey][key];
+                    Object.defineProperty(newSettings[(sectionKey as 'player' | 'workspace')], key, { value: settedValue });
+                }
+            }
+        }
+
+        newSettings.imageFilters = [];
+        if ('imageFilters' in loadedSettings) {
+            for (const filter of loadedSettings.imageFilters) {
+                switch (filter.alias) {
+                    case ImageFilterAlias.GAMMA_CORRECTION: {
+                        const modifier = new GammaCorrection(filter.params as GammaFilterOptions);
+                        newSettings.imageFilters.push({
+                            modifier,
+                            alias: ImageFilterAlias.GAMMA_CORRECTION,
+                        });
+                        break;
+                    }
+                    default: {
+                        break;
+                    }
+                }
+            }
+        }
+
+        dispatch(setSettings(newSettings));
+        if ('shortcuts' in loadedSettings) {
+            const updateKeyMap = structuredClone(shortcuts.keyMap);
+            for (const key of Object.keys(loadedSettings.shortcuts.keyMap)) {
+                const value = loadedSettings.shortcuts.keyMap[key];
+                if (key in updateKeyMap) {
+                    updateKeyMap[key].sequences = value.sequences;
+                }
+            }
+            for (const key of Object.keys(updateKeyMap)) {
+                const currValue = {
+                    [key]: { ...updateKeyMap[key] },
+                };
+                const conflictingShortcuts = conflictDetector(currValue, shortcuts.keyMap);
+                if (conflictingShortcuts) {
+                    for (const conflictingShortcut of Object.keys(conflictingShortcuts)) {
+                        for (const sequence of currValue[key].sequences) {
+                            for (const conflictingSequence of conflictingShortcuts[conflictingShortcut].sequences) {
+                                if (conflict(sequence, conflictingSequence)) {
+                                    updateKeyMap[conflictingShortcut].sequences = [
+                                        ...updateKeyMap[conflictingShortcut].sequences.filter(
+                                            (s: string) => s !== conflictingSequence,
+                                        ),
+                                    ];
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            dispatch(shortcutsActions.registerShortcuts(updateKeyMap));
+        }
+    };
+}
+
+export function updateCachedSettings(settings: CombinedState['settings'], shortcuts: CombinedState['shortcuts']) {
+    const settingsForSaving: any = {
+        shortcuts: {
+            keyMap: {},
+        },
+        imageFilters: [],
+    };
+    const supportedImageFilters = [ImageFilterAlias.GAMMA_CORRECTION];
+    for (const [key, value] of Object.entries(settings)) {
+        if (['player', 'workspace'].includes(key)) {
+            settingsForSaving[key] = value;
+        }
+        if (key === 'imageFilters') {
+            const filters = [];
+            for (const filter of value) {
+                if (supportedImageFilters.includes(filter.alias)) {
+                    filters.push(filter.modifier.toJSON());
+                }
+            }
+            settingsForSaving.imageFilters = filters;
+        }
+    }
+    for (const [key] of Object.entries(shortcuts.keyMap)) {
+        if (key in shortcuts.defaultState) {
+            settingsForSaving.shortcuts.keyMap[key] = {
+                sequences: shortcuts.keyMap[key].sequences,
+            };
+        }
+    }
+
+    localStorage.setItem('clientSettings', JSON.stringify(settingsForSaving));
 }
